Default to Recognizing ' as a character delimiter
If ' is used to quote another symbol, then ' should be part of the
symbol itself, hence treating ' as an expression prefix. This is the
same thing the various Lisp modes do, and what scala-mode defaults to
with the "Expression prefixes" ("'") syntax-class. scala-mode later
fixes-up incorrect default cases in the case the item IS ACTUALLY a
character, by using syntax-propertize-function.

scala-ts-mode used the incorrect "Character quotes" syntax-class "'".
This meant that ' shoulde treated as an escape character and that the
immediate next character should lose its meaning. The only use for
this syntax-class is the \ symbol in TeX modes, because \ makes the
next TeX symbol lose its normal meaning.

I will treat Scala's ' symbol as just a "String quotes"
syntax-class ("\""), because that should give us a nice default state,
and in the case Scala's ' is used to quote a symbol in Scala code,
tree-sitter-scala should have detected it and treated it as a separate
case. This default decision may be incorrect in the long-run, but it
fixes the issues I have ad where parens are mismatched because of a '
in front of a parenthetical structure without whitespace separating
them. An example of what used to be a problematic use is given below.

(x - 'A') // Used to treat 'A and ') as escaped characters, like \A
and \) in TeX, which caused incorrect behaviors regarding next-sexp,
and highlighting.